Update files to place classes in the classes folder and refactor associated imports for correct file path.
