--- src/SomeClassYouChangeTheName.js
@@ -1,6 +1,0 @@
-class SomeClassYouChangeTheName {
-  // This is here only so you can see how a class is imported into the scripts.js file
-  // Rename it and structure it however you need in the project
-}
-
-export default SomeClassYouChangeTheName;
