+++ src/classes/UserRepository.js
@@ -17,4 +17,4 @@ class UserRepository {
   };
 };
 
-module.exports = UserRepository;
+export default UserRepository;
